% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dog_pubs}
\alias{dog_pubs}
\title{Dog publications}
\format{A data frame with 30 rows and 3 variables:
\describe{
  \item{doi}{The publication's DOI}
  \item{title}{The publication's title}
  \item{abstract}{The publication's abstract}
}}
\usage{
dog_pubs
}
\description{
A data frame containing PLOS publication data for publications related to
dogs. The purpose of this data frame is to provide an example of some text
to extract keywords from.
}
\keyword{datasets}
