% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos-tags.R
\docType{data}
\name{pos_tags}
\alias{pos_tags}
\title{Part-of-speech (POS) tags}
\format{An object of class \code{data.frame} with 36 rows and 2 columns.}
\usage{
pos_tags
}
\description{
A data frame containing all possible parts-of-speech, as per the
\href{https://cran.r-project.org/web/packages/openNLP/openNLP.pdf}{openNLP}
package. This list was taken from \href{http://martinschweinberger.de/docs/articles/PosTagR.pdf}{Part-Of-Speech Tagging with R}... \code{pos_tags} contains the
following two columns:
\describe{
  \item{tag}{The abbreviation for the part-of-speech (i.e., its tag)}
  \item{dscrp}{A short description of the part-of-speech}
}
}
\keyword{datasets}
