% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind-rakelist.R
\name{rbind_rakelist}
\alias{rbind_rakelist}
\title{rbind a rakelist}
\usage{
rbind_rakelist(rakelist, doc_id = NULL)
}
\arguments{
\item{rakelist}{A object of class \code{rakelist}, which you create by calling \code{\link{slowrake}}.}

\item{doc_id}{An optional vector of document IDs, which should be the same length as \code{rakelist}. These IDs will be added to results data frame.}
}
\description{
rbind a rakelist
}
