% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slowrake.R
\name{slowrake}
\alias{slowrake}
\title{Slow RAKE}
\usage{
slowrake(txt, stop_words = smart_words, word_min_char = 3, stem = TRUE,
  keep_pos = c("NN", "NNS", "NNP", "NNPS"))
}
\arguments{
\item{txt}{A character vector, where each element of the vector contains the text for one document.}

\item{stop_words}{A vector of stop words, which will be removed from your corpus prior to the keyword extraction algorithm being applied. The default value (\code{smart_words}) is the "SMART" vector of stop words (equivilent to \code{\link[tm]{stopwords("SMART")}}).}

\item{word_min_char}{The minimum number of characters that a word must have to remain in the corpus. Words with fewer than \code{word_min_char} characters will be removed prior before keyword extraction is conducted.}

\item{stem}{Do you want to stem the words in your corpus before extracting keywords?}

\item{keep_pos}{A vector of part-of-speech (POS) tags that specifies accepitble parts-of-speech for keywords to have. The default value \code{c("NN", "NNS", "NNP", "NNPS")} specifies that all words that have a POS that is noun-related (e.g., plural noun, proper noun, etc.) will be considered when making the candidate keywords. To see the other possible POS tags along with their descriptions, see the \code{\link{pos_tags}} data frame (\code{View(slowraker::pos_tags)}) or see their descriptions in \href{Part-Of-Speech Tagging with R}{http://martinschweinberger.de/docs/articles/PosTagR.pdf}. Specifying \code{keep_pos = NULL} will not filter out any words based on their POS.}
}
\value{
A list of data frames, with one data frame per document. Each data frame will have the following columns:
 \describe{
   \item{keyword}{A keyword that was identified by RAKE. This list is non-distinct, meaning that a keyword can appear in this column nice if it appears in the text twice.}

   \item{freq}{The number of times the keyword appears in the document}

   \item{score}{The keyword's score, as per the RAKE algorithm. Keywords with higher scores are considered to be more higher quality than those with lower scores.}

   \itemP{stem}{If you specified \code{stem = TRUE}, you will also get the stemmed versions of the keywords in this column. When you choose stemming, the keyword's score (\code{score}) will be based off of its stem, but the reported number of times that the keyword appears (\code{freq}) will still be based off of the raw, unstemmed variant of the keyword.}
 }
}
\description{
Slow RAKE
}
