% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart-words.R
\docType{data}
\name{smart_words}
\alias{smart_words}
\title{SMART stop words}
\format{An object of class \code{character} of length 571.}
\usage{
smart_words
}
\description{
A vector containing the SMART information retrieval system stop words. See
\href{https://rdrr.io/rforge/tm/man/stopwords.html}{tm::stopwords('SMART')}
for more details.
}
\keyword{datasets}
